Add and implement all loader classes
