+++ include/Map.h
@@ -32,6 +32,8 @@ class Map
 
         friend class TMXParser;
         friend class MapLoader;
+        friend class TilesetLoader;
+        friend class LayerLoader;
         friend class MapRenderer;
 
     private:
